etc: prompt color order and fixes
* bashrc sets the prompt and exports the prompt color to userrc. userrc
may set the prompt as it wishes or leave it untouched.
* colors become a mechanism. they may have various uses, so the colors
only code the lowest common demoninator. \E or \033 are added by the
user. to do this gradually, the new (minimal) codes are named _COLOR and
the previous __COLOR remain. when all code is adjusted, _COLOR will
become __COLOR.
* __set_prompt only sets settings, does not dereference colors. all
dereferencing is done by PS1 itself.

Signed-off-by: Idan Kedar <idank@tonian.com>