bashrc: no user name in prompt (optional)
allows the user to configure whether or not PS1 displays the user name.
a typical work process is either being your own user, or being root, and
the prompt color already gives the user a fair indication of that.

the default is to display the user name, as it was until now.

Signed-off-by: Idan Kedar <idank@tonian.com>