bash: complete user-defined scripts for root commands
consider the following case:
$ ll
-rwxr--r--  1 root data 1.3K Feb  3  2012 import.sh*

this is a script that is executable by root, but will not be
auto-completed when you try:

$ sudo ./imp<Tab>

this patch makes it completable. note that this does not apply only for
scripts, but for everything executable.

on the down side, this patch causes scripts executable by the user to be
excluded from completion, but i consider this OK since i don't think
it's likely to want to run scripts belonging to a user with sudo.

Signed-off-by: Idan Kedar <idank@tonian.com>