git-make-patchset: add a back reference from patch to cover letter
for situations where one has a patch from a patchset, and needs a
broader context (the patchset), the patchset name and identifier will
allow a decent search engine to find all related patches.

the patchset ID consists of the date, version, under the assumption that
no two people will generate a patchset with the same name and version on
the same second.

i wanted to avoid situations where a patch may have the ID bcd and might
match patches abcd, bcde or abcde. because the date is unique and of a
fixed string length, the abcd case shouldn't happen at all, and the
abcde case is equivalent to the bcde case.
the "bcd matching bcde" case is actually a "v1 matches v14" case.
to avoid it, i chose a random letter ('T'), and appended it to the ID.
this way, bcd becomes bcdT, bcde becomes bcdeT, and as such, bcdT is not
a substring of bcdeT and will not match it.

Signed-off-by: Idan Kedar <idank@primarydata.com>