git-make-pachset: set text width to 80 when the editor is vi
...or vim.
bash refuses to accept VISUAL_FLAGS as a string - it kept doing word
splitting no matter what quoting and escaping i tried. so i had to
resort to an array.

the alternative was switching on $VISUAL in the call, which is verbose,
or editing the file with eval, which i wanted to avoid since eval is
always hard to get.

Signed-off-by: Idan Kedar <idank@primarydata.com>