introduce rbt-post-patchset
* post and publish the first review request
* post and publish each subsequent patch as an update
* let people know where the branch is pushed
* insert the git log (commit messages) with per-patch diffstat to the
description

the script assumes you have your username configured, but not the
password, so it will ask for the password at the beginning and then
issue rbt commands with --password.

the biggest issue currently is that i could not find a way to post an
update without publishing the previous update, so if the script
doesn't publish, the updates are not seen as such.

Signed-off-by: Idan Kedar <idank@primarydata.com>