introduce pristine/ directory
in some cases, e.g. c.vim, we're overwriting a whole file, when we
actually just want to change one thing.

we could make a patch and apply it, and maybe we will in the future, but
for now we just keep the pristine file that we're overwriting, so that
if something goes wrong we can diff the one in pristine/ vs its
counterpart in src/ and figure out what we meant to change so that we
can fix things in case a patch doesn't work.

Signed-off-by: Idan Kedar <idank@primarydata.com>