urxvt: even better keyboard handling
so apparently in urxvt 9.20 the "clipboard" extension was presumably
rendered obsolete by urxvt's built-in support for autocopy, in
addition to selection-to-clipboard, which apparently is a thing.

the only reason to still use the "clipboard" extension is the following
sequence:
1. ctrl+C something in any window (copied to CLIPBOARD)
2. double-click something in any window (copied to PRIMARY)
3. switch to urxvt
4. paste

ideally, that would paste from CLIPBOARD, but unfortunately that pastes
from PRIMARY - unless you configure paste to paste from clipboard.

unfortunately that only works for shift+insert and not middle click.

Signed-off-by: Idan Kedar <ik.imbaf@gmail.com>