introducing httpc
like errno and signum, but for http status codes

all the status code databases seem to be online rather than available in
packages, so the database has to be committed as well.

note that it's grepping the entire verbose message, even if that message
isn't displayed.

it's taken from https://httpstatuses.com by copying the list in that
page, then:
$ : > /tmp/statuses
$ CODES=$(awk '{ print $1 }' codes_file.txt)
$ for S in $CODES; do
	curl https://httpstatuses.com/$S |
	grep 'description.*\.">' |
	grep -o ": .*\." >> /tmp/statuses
done
$ sed -i 's/&quot;/"/' /tmp/statuses
$ magic
$ paste -d "" codes_file.txt /tmp/statuses > verbose_file.txt

the "magic" part is actually a manual step because for some reason the
curl command didn't work for status code 202 (Accepted) and I had to add
it manually to /tmp/statuses before moving to the paste step.

in addition, the magic phase includes manual things like removing
unnecessary text from /tmp/statuses.

Signed-off-by: Idan Kedar <idank@imperva.com>