make: disable caches
desktop applications are terrible at security.
in particular, they cache stuff on-disk for any other application run by
that user to read.
usually that's a good thing.
when you don't trust your own applications, that isn't a good thing.

almost no application will go as far as deleting a user's file, so this
is our method of choice for disabling caches: where a directory is
expected, we put a file with data; where a file is expected, we put a
directory with a single file with data.

some applications have been seen crashing when such unexpected cache
behavior occurs. that's why the caches are fully configurable using
CONFIG_DISABLE_CACHE_FILES and CONFIG_DISABLE_CACHE_DIRS.

Signed-off-by: Idan Kedar <ik.imbaf@gmail.com>