make: desktopinit
that's a target you run once after you installed your desktop.
the main purpose is installing web-browser profiles and such.

since those are quite large (in the megabytes), they are placed in a
separate github repo:
  https://github.com/saint-carlos/desktopinit.git/

so when you type
  $ make install_desktopinit

it clones that repo and installs the profiles which aren't already
configured.

if you want to kill your old profiles in order to make this work:
  $ rm -r $HOME/.mozilla/firefox $HOME/.config/chromium

Signed-off-by: Idan Kedar <ik.imbaf@gmail.com>