make: optimize the configuration process
obtaining the value of CONFIG_DCONF_LOCATIONS was apparently super slow.
now that it's optimized, build has gone from 37 seconds to 2 seconds.

the way to optimize it is to make the final configuration and the sed
script be files and not regenerate them every time.

it's somewhat intrusive, but a 19-fold improvement is worth it.

Signed-off-by: Idan Kedar <ik.imbaf@gmail.com>