much better touchpad support
first, we adjust the touchpad settings themselves to be more like Mac,
not because Mac is so great, but because people are used to Mac
touchpads more than they're used to Linux touchpads (not talking about
Windows here...).

additionally in a touchpad keyboard, there's no numeric keypad, so some
shortcuts should also be changed.
we put these shortcuts in different files, and they differ based on the
keyboard type.

at this point we have quite a lot of files, so they get their own
directory.

Signed-off-by: Idan Kedar <ik.imbaf@gmail.com>