git: much better paging
the generic --pattern is bad when the pattern is not found and/or when
the result is less than a page.

but with some per-command tuning it's much better.

commands where we expect paging now run less with -c instead -f -FX,
and we also start at the beginning with '+g'.

Signed-off-by: Idan Kedar <ik.imbaf@gmail.com>