vimrc: map Alt+j and Alt+k to gj and gk
the main issue is that urxvt is unicode, so something like Alt+j marks
the beginning of a utf8 character input rather then sending an Alt'ed
j (0x6a + 8th bit = 0xea) to the TTY.

at least that's the behavior with the urxvt "meta8" option, which
defaults to 'false'.

when it is false, urxvt just sends ^[j (to generate ^[ press
Ctrl+V Ctrl+[), so vim doesn't see <M-j>, it sees ^[j. therefore, we
need to map ^[j to <M-j> etc.

weirdly enough, setting urxvt's "meta8" option to true does allow
inserting unicode character, neither 0xea nor ^[j, because
it's expecting a unicode character. it does work if we add to .vimrc:
  set termencoding=latin1
which tells vim that the terminal is NOT unicode set the TTY to just
send the 0xea.

unfortunately, that also makes vim not display unicode characters at
all, and adding
  set encoding=utf-8
doesn't help.

so for now we're stuck with manual mapping.

more insights here:
https://vim.fandom.com/wiki/Get_Alt_key_to_work_in_terminal

Signed-off-by: Idan Kedar <ik.imbaf@gmail.com>