vim: fix the super-annoying 1-second delay when exiting visual mode
how to reproduce:
* enter visual mode
* press <Esc>

expected behavior:
* normal mode

actual behavior:
* 1 second delay, waiting for a key to follow <Esc>
* normal mode after the delay

triage:
fucking TabBar again. it's has visual mode mappings that start with
<Esc>.
but even if it doesn't, vim still waits for key codes.

solution:
the internet says to set ttimeoutlen=0.
despite what :h InsertEnter says, it is NOT triggered for visual mode,
so we set it outside. that takes care of key codes.
but we still ned to take care of TabBar's mapped keys.
so we have TabBar not create visual mode mappings, we don't need to
switch tabs outside of normal mode.

Signed-off-by: Idan Kedar <ik.imbaf@gmail.com>