gitconfig: fix no-pager for aliases *AGAIN*
"hstat" screws around with the terminal, and that doesn't get
automatically fixed.

when the pager is "less" or "wless", this one-line output gets paged,
which is not good.

when the pager is "less -FXR" or wless "-FXR", this one doesn't get
paged but it may or may not screw up the terminal based on the the
value of pager.hstat.

if pager.hstat is "true" or "less -FXR" or "wless -FXR", it goes though
"less -FXR", which fixes things up.
if pager.hstat is "false", or "cat -", then "cat -" is used, and no
terminal fixup takes place.
if pager.hstat is "cat", then git ignores the pager for this alias and
uses the pager of the underlying command, which is "wless -cR"

Signed-off-by: Idan Kedar <ik.imbaf@gmail.com>