bashrc: more useful timestamps
instead of printing the time the command finished in the prompt, print
the time it started + how long it took, which is generally more useful.
if needed, that can be used to compute when it ended.

prefix it with '^' (up arrow) to remind that it's from the previous
command.

this relies on the fancy PS0, which runs before a command is executed,
and on the variable dereference syntax that allows assigning (possibly
global) variables, see https://stackoverflow.com/a/66772796.

Signed-off-by: Idan Kedar <ik.imbaf@gmail.com>