make: record empty backups
if a config file, e.g. gitconfig, was not present, it was being
installed with no backup. therefore when uninstalled, since there was no
backup, the file was assumed to not have been configured, and therefore
not uninstalled.

so we record the fact that the file does not exist.

it is possible that the file exists but empty, and that the respective
app distinguishes missing file from empty file, so we explicitly record
missing files.

TODO probably better to have a manifest of installed files.

Signed-off-by: Idan Kedar <ik.imbaf@gmail.com>