make: 2-stage makefile
the makefile itself depends on the configuration. this makes for a very
confusing build.

split the build to 2 phases:
the toplevel makefile sets up all the configuration, and delegates the
actual build to build/makefile.
build/makefile performs the actual build & install, assuming that the
configuration has been set up correctly.

Signed-off-by: Idan Kedar <ik.imbaf@gmail.com>